optimizations for 3.0.0

+++ app/Log.php
@@ -8,7 +8,7 @@
 namespace PersonioIntegrationLight;
 
 // prevent direct access.
-defined( 'ABSPATH' ) or exit;
+defined( 'ABSPATH' ) || exit;
 
 /**
  * Handler for logging in this plugin.
@@ -78,6 +78,11 @@ class Log {
 	 * @return void
 	 */
 	public function clean_log(): void {
+		// bail on uninstalling.
+		if ( defined( 'PERSONIO_INTEGRATION_DEACTIVATION_RUNNING' ) ) {
+			return;
+		}
+
 		global $wpdb;
 		$wpdb->query( sprintf( 'DELETE FROM %s WHERE `time` < DATE_SUB(NOW(), INTERVAL %d DAY)', esc_sql( $wpdb->prefix . 'personio_import_logs' ), absint( get_option( 'personioIntegrationMaxAgeLogEntries' ) ) ) );
 	}
